refactor: 사용하지 않는 인터페이스 제거
